requirements.txt
requests
beautifulsoup4
pandas
groq
python-dotenv